Rorganized all routes and controller methods working with invites to be organized in the invites controller and route files.
