+++ app/routes/invites.js
@@ -1,12 +1,25 @@
 const router = require("express").Router();
 const isAuthenticated = require("./middleware/isAuthenticated");
 const setInvite = require("./middleware/setInvite");
+const setRoom = require("./middleware/setRoom");
+const isRoomMember = require("./middleware/isRoomMember");
 
 const {
+    create,
+    remove,
     accept,
     emailCheck,
     register
 } = require("../controllers/invite");
+
+
+router
+    .route('/:roomId')
+    .post( isAuthenticated, setRoom.fromParam, isRoomMember, create );
+
+router
+    .route('/:roomId/:inviteId')
+    .delete( isAuthenticated, setRoom.fromParam, isRoomMember, remove );
 
 router
     .route('/:token')
