+++ app/routes/isAuth/rooms.js
@@ -5,22 +5,12 @@ const isRoomMember = require("../middleware/isRoomMember");
 
 const {
     getSingle,
-    update,
-    createInvite,
-    deleteInvite
+    update
 } = require("../../controllers/room");
 
 router
     .route( "/:roomId" )
     .get( setRoom.fromParam, isRoomMember, getSingle )
     .patch( setRoom.fromParam, isRoomMember, update );
-
-router
-    .route( "/:roomId/invites" )
-    .post( setRoom.fromParam, isRoomMember, createInvite )
-
-router
-    .route( "/:roomId/invites/:inviteId" )
-    .delete( setRoom.fromParam, isRoomMember, deleteInvite )
 
 module.exports = router;
